+++ app/src/main/java/cord/eoeo/momentwo/ui/albumdetail/member/MemberListItem.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import cord.eoeo.momentwo.ui.model.MemberAuth
 import cord.eoeo.momentwo.ui.model.MemberItem
-import cord.eoeo.momentwo.ui.theme.starYellow
+import cord.eoeo.momentwo.core.designsystem.theme.starYellow
 
 @Composable
 fun MemberListItem(
