+++ app/src/main/java/cord/eoeo/momentwo/ui/composable/SearchUserDialog.kt
@@ -46,7 +46,7 @@ import androidx.paging.compose.itemKey
 import coil.ImageLoader
 import coil.imageLoader
 import cord.eoeo.momentwo.ui.model.UserItem
-import cord.eoeo.momentwo.ui.theme.MomentwoTheme
+import cord.eoeo.momentwo.core.designsystem.theme.MomentwoTheme
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
