+++ app/src/main/java/cord/eoeo/momentwo/ui/photolist/PhotoListItemBox.kt
@@ -25,8 +25,8 @@ import androidx.compose.ui.unit.dp
 import coil.ImageLoader
 import coil.compose.AsyncImage
 import cord.eoeo.momentwo.ui.model.PhotoItem
-import cord.eoeo.momentwo.ui.theme.favoriteBorder
-import cord.eoeo.momentwo.ui.theme.primaryDark
+import cord.eoeo.momentwo.core.designsystem.theme.favoriteBorder
+import cord.eoeo.momentwo.core.designsystem.theme.primaryDark
 
 @Composable
 fun PhotoListItemBox(
