+++ core/designsystem/src/main/java/cord/eoeo/momentwo/core/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package cord.eoeo.momentwo.ui.theme
+package cord.eoeo.momentwo.core.designsystem.theme
 
 import androidx.compose.material3.Typography
